Palmistry
The Project titled “Palm Recognition System” with the aim of
implementing key concepts in
Data Mining was found to meet all objectives. Utmost care has been taken
to see that the results
obtained are true and the underlying procedures are implemented in true
form. The aim of this
project is to present a novel software based selection method that can
be used by anyone to get
predictions by scanning only the image of their hands. It provides a
User-Friendly way to select
the lines they want to interpret for future predictions. We have used as
our basic and core
algorithm for computing and predicting the result. The proposed method
will provide a novel
way to read lines in an affective budget friendly way.